hrl/agent/bonus_based_exploration/requirements.txt
absl-py
atari_py
dopamine-rl
gin-config >= 0.1
# matplotlib
numpy >= 1.14
pillow
pygame
six
scipy
tensorflow
